Added files via upload
This is my nearly finished version of the map tool. I'm really bad at this github thing so I didn't know how to update the existing one. The last thing that needs to be added to this is the ability to save the map, which I've started but I can't figure it out.